split into lib and main
